Removed unused imports, added @Override tags, ...
Code cleanup

+++ org.jcryptool.analysis.graphtools/src/org/jcryptool/analysis/graphtools/derivates/OverlayBar.java
@@ -87,7 +87,7 @@ public class OverlayBar extends Bar {
 	public void drawBar(Rectangle myBox, double myHeight, double myWidth, MColor myColorMainbar, GC myGC) {
 		int barWidth = (int) Math.round(myWidth * myBox.width);
 		int barHeight = (int) Math.round(myHeight * myBox.height);
-		int topleftX = (int) Math.round((double) (myBox.width - myWidth * myBox.width) / 2 + myBox.x);
+		int topleftX = (int) Math.round((myBox.width - myWidth * myBox.width) / 2 + myBox.x);
 		int topleftY = myBox.y + myBox.height - barHeight;
 		int transparencyShades = 20;
 		int transparencySteps = Math.max(1, (barHeight / transparencyShades));
